LayoutTests/platform/mac-wk1/http/tests/navigation/sec-fetch-site-header-expected.txt

Tests the behavior of sec-fetch-site header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL secFetchSite should be none. Was cross-site.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE


============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/sec-fetch-site-header.html
curr->  https://localhost:8443/referrer-policy/resources/check-sec-fetch-site.py?expected=none
===============================================
